refactor(docs): extract and test resolveXuluxModel
